+++ src/test/suite/kafka-file/languageservice/completionProperties.test.ts
@@ -567,7 +567,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 0))
                     },
                     {
@@ -608,7 +608,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 1))
                     },
                     {
@@ -648,7 +648,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 5))
                     },
                     {
@@ -688,7 +688,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
                     },
                     {
@@ -728,7 +728,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
                     },
                     {
@@ -768,7 +768,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 11))
                     },
                     {
@@ -810,7 +810,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
                         range: range(position(2, 0), position(2, 5))
                     },
                     {
@@ -1019,6 +1019,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     {
                         label: 'string with encoding...', kind: CompletionItemKind.Value,
                         insertText: ' string(${1|utf8,utf16le,base64,latin1,hex|})',
+                        range: range(position(1, 13), position(1, 13))
+                    },
+                    {
+                        label: 'json', kind: CompletionItemKind.Value,
+                        insertText: ' json',
                         range: range(position(1, 13), position(1, 13))
                     },
                     {